+++ Assets/Script/CatArmour.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class CatArmour : Enemy {
     
-    public CatArmour (int baseAttack, int attackModifier, int baseShield, int shieldModifier) 
+    public CatArmour (int baseAttack, double attackModifier, int baseShield, double shieldModifier) 
     : base(30, baseAttack, attackModifier, baseShield, shieldModifier) { 
         //empty
     }
