+++ Assets/Script/CatStaff.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class CatStaff : Enemy {
-    public CatStaff(int baseAttack, int attackModifier, int baseShield, int shieldModifier)
+    public CatStaff(int baseAttack, double attackModifier, int baseShield, double shieldModifier)
     : base(30, baseAttack, attackModifier, baseShield, shieldModifier) {
         
     }
