+++ Assets/Script/CatSword.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class CatSword : Enemy {
     
-    public CatSword (int baseAttack, int attackModifier, int baseShield, int shieldModifier) 
+    public CatSword (int baseAttack, double attackModifier, int baseShield, double shieldModifier) 
     : base(30, baseAttack, attackModifier, baseShield, shieldModifier) { 
         //empty
     }
