+++ Assets/Script/Mage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class Mage : Player {
 
 
-    public Mage(int baseAttack, int attackModifier, int baseShield, int shieldModifier) 
+    public Mage(int baseAttack, double attackModifier, int baseShield, double shieldModifier) 
     : base(80, baseAttack, attackModifier, baseShield, shieldModifier) { 
         //empty
     }
